+++ CHANGELOG.md
@@ -158,3 +158,4 @@
 - [2026-04-06 12:23] feat(ui): implement responsive ui view component
 - [2026-04-06 12:32] feat(storage): add bi-directional event bus handler for storage
 - [2026-04-06 13:17] feat(i18n): implement responsive i18n view component
+- [2026-04-06 14:14] style(state): harmonize light and dark theme palette for state